+++ EnhanceQoL/Submodules/ChatIM/UI.lua
@@ -9,6 +9,7 @@ end
 
 local L = LibStub("AceLocale-3.0"):GetLocale("EnhanceQoL")
 local canaccessvalue = _G.canaccessvalue
+local MouseIsOver = addon.functions.MouseIsOver
 
 local function colorWrap(hex, text) return "|cff" .. hex .. text .. "|r" end